+++ packages/raw-data-provider/lib/providers/cardano/index.ts
@@ -1,1 +1,2 @@
 export { CardanoKoiosRawDataProvider } from './cardanoKoiosRawDataProvider';
+export { CardanoOgmiosRawDataProvider } from './cardanoOgmiosRawDataProvider';