+++ packages/raw-data-provider/tests/utils.ts
@@ -2,6 +2,10 @@ import {
   ObservationEntity,
   migrations as observationMigrations,
 } from '@rosen-bridge/abstract-observation-extractor';
+import {
+  BlockEntity,
+  migrations as blockMigrations,
+} from '@rosen-bridge/abstract-scanner';
 import { DataSource } from '@rosen-bridge/extended-typeorm';
 
 import { RawDataProviderStateEntity } from '../lib/entities';
@@ -11,8 +15,12 @@ export const createDatabase = async (): Promise<DataSource> => {
   return new DataSource({
     type: 'sqlite',
     database: `:memory:`,
-    entities: [RawDataProviderStateEntity, ObservationEntity],
-    migrations: [...migrations.sqlite, ...observationMigrations.sqlite],
+    entities: [BlockEntity, RawDataProviderStateEntity, ObservationEntity],
+    migrations: [
+      ...blockMigrations.sqlite,
+      ...migrations.sqlite,
+      ...observationMigrations.sqlite,
+    ],
     synchronize: false,
     logging: false,
   })