+++ packages/observation-extractors/bitcoin-observation-extractor/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @rosen-bridge/bitcoin-observation-extractor
 
+## 7.0.1
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@1.0.1
+  - @rosen-bridge/bitcoin-scanner@1.0.1
+  - @rosen-bridge/rosen-extractor@11.2.1
+  - @rosen-bridge/tokens@5.0.1
+
+## 7.0.0
+
+### Major Changes
+
+- Update AbstractExtractor interface; use `BlockInfo` in `processTransactions` and rename `initializeBoxes` to `initializeData`
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-logger@4.0.0
+  - @rosen-bridge/abstract-observation-extractor@1.0.0
+  - @rosen-bridge/bitcoin-scanner@1.0.0
+  - @rosen-bridge/rosen-extractor@11.2.0
+  - @rosen-bridge/tokens@5.0.0
+
 ## 6.4.1
 
 ### Patch Changes